Broke out output state parsing into separate file.
